templates/common/backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (Async SQLAlchemy 2.0)
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and utilities
python-dotenv==1.0.0
httpx==0.25.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0

# Additional utilities
uuid6==2023.5.2
python-slugify==8.0.1